move HumanEntry and HeartEmoji to Models folder
