+++ freewrite/Models/HumanEntry.swift
@@ -1,0 +1,34 @@
+//
+//  HumanEntry.swift
+//  freewrite
+//
+//  Created by Gaspar Dolcemascolo on 17-04-25.
+//
+
+import Foundation
+
+struct HumanEntry: Identifiable {
+    let id: UUID
+    let date: String
+    let filename: String
+    var previewText: String
+    
+    static func createNew() -> HumanEntry {
+        let id = UUID()
+        let now = Date()
+        let dateFormatter = DateFormatter()
+        dateFormatter.dateFormat = "yyyy-MM-dd-HH-mm-ss"
+        let dateString = dateFormatter.string(from: now)
+        
+        // For display
+        dateFormatter.dateFormat = "MMM d"
+        let displayDate = dateFormatter.string(from: now)
+        
+        return HumanEntry(
+            id: id,
+            date: displayDate,
+            filename: "[\(id)]-[\(dateString)].md",
+            previewText: ""
+        )
+    }
+}
